wrap the resource in a class too
